Menu HUD + Item disappear + End board + New Items
The game now has a title screen.
The game now has a menu hud to choose the number of players.
The items have a live duration, once the time's up, the item disappear
if not picked.
At the end of a game, an end board displays the score and classement of
the players.
A life item and Invert control Item types have been added.